ciao-launcher: Move hard-reset code to own file
This commit consolidates all of the code for hard-reset
into one file.  Previously the code was spread throughout launcher
which wasn't ideal, particularly as the functions dockerKillInstance
and qemuKillInstance were stored in files reserved for the instance
go routine, even though these functions are called by the main
go routine.

Fixes #3
